--- a/bookkeeper-server/src/main/java/org/apache/bookkeeper/bookie/storage/ldb/ArrayGroupSort.java
+++ b/bookkeeper-server/src/main/java/org/apache/bookkeeper/bookie/storage/ldb/ArrayGroupSort.java
@@ -22,58 +22,61 @@ package org.apache.bookkeeper.bookie.storage.ldb;
 
 import static com.google.common.base.Preconditions.checkArgument;
 
+import lombok.experimental.UtilityClass;
+
 /**
  * Sort an array of longs, grouping the items in tuples.
  *
  * <p>Group size decides how many longs are included in the tuples and key size controls how many items to use for
  * comparison.
  */
+@UtilityClass
 public class ArrayGroupSort {
 
-    private final int keySize;
-    private final int groupSize;
+    private static final int INSERTION_SORT_THRESHOLD = 100;
 
-    public ArrayGroupSort(int keySize, int groupSize) {
-        checkArgument(keySize > 0);
-        checkArgument(groupSize > 0);
-        checkArgument(keySize <= groupSize, "keySize need to be less or equal the groupSize");
-        this.keySize = keySize;
-        this.groupSize = groupSize;
-    }
+    private static final int GROUP_SIZE = 4;
 
     public void sort(long[] array) {
         sort(array, 0, array.length);
     }
 
-    public void sort(long[] array, int offset, int length) {
-        checkArgument(length % groupSize == 0, "Array length must be multiple of groupSize");
-        quickSort(array, offset, (length + offset - groupSize));
+    public static void sort(long[] array, int offset, int length) {
+        checkArgument(length % GROUP_SIZE == 0, "Array length must be multiple of 4");
+        quickSort(array, offset, (length + offset - GROUP_SIZE));
     }
 
     ////// Private
 
-    private void quickSort(long[] array, int low, int high) {
-        if (low < high) {
-            int pivotIdx = partition(array, low, high);
-            quickSort(array, low, pivotIdx - groupSize);
-            quickSort(array, pivotIdx + groupSize, high);
+    private static void quickSort(long[] array, int low, int high) {
+        if (low >= high) {
+            return;
         }
+
+        if (high - low < INSERTION_SORT_THRESHOLD) {
+            insertionSort(array, low, high);
+            return;
+        }
+
+        int pivotIdx = partition(array, low, high);
+        quickSort(array, low, pivotIdx - GROUP_SIZE);
+        quickSort(array, pivotIdx + GROUP_SIZE, high);
     }
 
-    private int alignGroup(int count) {
-        return count - (count % groupSize);
+    private static int alignGroup(int count) {
+        return count - (count % GROUP_SIZE);
     }
 
-    private int partition(long[] array, int low, int high) {
+    private static int partition(long[] array, int low, int high) {
         int mid = low + alignGroup((high - low) / 2);
         swap(array, mid, high);
 
         int i = low;
 
-        for (int j = low; j < high; j += groupSize) {
+        for (int j = low; j < high; j += GROUP_SIZE) {
             if (isLess(array, j, high)) {
                 swap(array, j, i);
-                i += groupSize;
+                i += GROUP_SIZE;
             }
         }
 
@@ -81,26 +84,43 @@ public class ArrayGroupSort {
         return i;
     }
 
-    private void swap(long[] array, int a, int b) {
-        long tmp;
-        for (int k = 0; k < groupSize; k++) {
-            tmp = array[a + k];
-            array[a + k] = array[b + k];
-            array[b + k] = tmp;
-        }
+    private static void swap(long[] array, int a, int b) {
+        long tmp0 = array[a];
+        long tmp1 = array[a + 1];
+        long tmp2 = array[a + 2];
+        long tmp3 = array[a + 3];
+
+        array[a] = array[b];
+        array[a + 1] = array[b + 1];
+        array[a + 2] = array[b + 2];
+        array[a + 3] = array[b + 3];
+
+        array[b] = tmp0;
+        array[b + 1] = tmp1;
+        array[b + 2] = tmp2;
+        array[b + 3] = tmp3;
     }
 
-    private boolean isLess(long[] array, int idx1, int idx2) {
-        for (int i = 0; i < keySize; i++) {
-            long k1 = array[idx1 + i];
-            long k2 = array[idx2 + i];
-            if (k1 < k2) {
-                return true;
-            } else if (k1 > k2) {
-                return false;
-            }
+    private static boolean isLess(long[] array, int a, int b) {
+        long a0 = array[a];
+        long b0 = array[b];
+
+        if (a0 < b0) {
+            return true;
+        } else if (a0 > b0) {
+            return false;
         }
 
-        return false;
+        return array[a + 1] < array[b + 1];
+    }
+
+    private static void insertionSort(long[] a, int low, int high) {
+        for (int i = low + GROUP_SIZE; i <= high; i += GROUP_SIZE) {
+            int j = i;
+            while (j > 0 && isLess(a, j, j - GROUP_SIZE)) {
+                swap(a, j, j - GROUP_SIZE);
+                j -= GROUP_SIZE;
+            }
+        }
     }
 }